+++ src/main/java/com/nongsa/dto/ReplySaveRequestDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReplySaveRequestDto {
-    private int userId;
-    private int boardId;
+    private Long userId;
+    private Long boardId;
     private String content;
 }
