+++ src/main/java/com/nongsa/repository/SubscribeRepository.java
@@ -5,22 +5,22 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface SubscribeRepository extends JpaRepository<Subscribe, Integer> {
+public interface SubscribeRepository extends JpaRepository<Subscribe, Long> {
 
     @Modifying
     @Query(value = "insert into subscribe(fromUserId,toUserId,createDate) values(:fromUserId,:toUserId,now())", nativeQuery = true)
-    void subscribe(int fromUserId, int toUserId);
+    void subscribe(Long fromUserId, Long toUserId);
 
     @Modifying
     @Query(value = "delete from subscribe where fromUserId=:fromUserId and toUserId=:toUserId", nativeQuery = true)
-    void unSubscribe(int fromUserId, int toUserId);
+    void unSubscribe(Long fromUserId, Long toUserId);
 
     @Query(value = "SELECT COUNT(*) FROM subscribe WHERE fromUserId = :principalId AND toUserId=:pageUserId", nativeQuery = true)
-    int subscribeState(int principalId, int pageUserId);
+    int subscribeState(Long principalId, Long pageUserId);
 
     @Query(value = "SELECT COUNT(*) FROM subscribe WHERE fromUserId =:pageUserId", nativeQuery = true)
-    int subscribeCount(int pageUserId);
+    int subscribeCount(Long pageUserId);
 
     @Query(value = "SELECT COUNT(*) FROM subscribe WHERE toUserId =:pageUserId", nativeQuery = true)
-    int subscribedCount(int pageUserId);
+    int subscribedCount(Long pageUserId);
 }
