+++ src/main/java/com/nongsa/service/LikesService.java
@@ -11,12 +11,12 @@ public class LikesService {
     private final LikesRepository likesRepository;
 
     @Transactional
-    public void 좋아요(int boardId, int principalId) {
+    public void 좋아요(Long boardId, Long principalId) {
         likesRepository.likes(boardId, principalId);
     }
 
     @Transactional
-    public void 좋아요취소(int boardId, int principalId) {
+    public void 좋아요취소(Long boardId, Long principalId) {
         likesRepository.unLikes(boardId, principalId);
     }
 }
